requirements.txt
# requirements.txt

chardet==5.2.0
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
XlsxWriter==3.2.0
